+++ fiddles/extjs6/fiddle-0000-Template/app.js
@@ -1,6 +1,6 @@
 var meta = {
-    fiddleHeader: 'fiddle-0032-CSVStore',
-    fiddleSubHeader: 'Template Fiddle created @ 01-22-16' +
+    fiddleHeader: 'fiddle-0000-Template',
+    fiddleSubHeader: 'Template Fiddle created @ 12-19-18' +
         '<br />'
 };