+++ fiddles/extjs6/fiddle-0014-WidgetGridKS/src/meta.js
@@ -1,9 +1,0 @@
-var meta = {
-    fiddleHeader: 'Widget Grid Kitchen Sink',
-    fiddleSubHeader: 'Fiddle lifted from the Kitchen Sink - Widget Grid.' +
-    '<br />'
-};
-
-var fiddleStore = null;
-
-