+++ fiddles/extjs6/fiddle-0023-ReporterSpecRunner/src/meta.js
@@ -1,5 +1,0 @@
-var meta = {
-    fiddleHeader: 'Reporter Tear Down',
-    fiddleSubHeader: 'Fiddle exploring how to establish associations between tree models' +
-    '<br />'
-};