+++ fiddles/extjs6/fiddle-0032-CSVStore/src/meta.js
@@ -1,9 +1,0 @@
-var meta = {
-    fiddleHeader: 'fiddle-0032-CSVStore',
-    fiddleSubHeader: 'Template Fiddle created @ 01-22-16' +
-    '<br />'
-};
-
-var fiddleStore = null;
-
-